ts_queue uses an internal pointer array instead of kqueue
Kqueue required data items stored in the queue to implement KListElem.
This however is not a desirable API if an alternate implementation is
provided independent of klist/kqueue. So I changed the ts_queue
implementation to use a pointer array to store client supplied pointer
items. The array is of fixed size and helper macros are provided to create
instances of ts_queue from static storage.